% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordlists.R
\name{get_sentiment_words}
\alias{get_sentiment_words}
\title{Create wordlists}
\usage{
get_sentiment_words()
}
\value{
A list with two elements: positive and negative words
}
\description{
Returns vectors of simple positive and negative sentiment words or stems,
based on observed patterns in the example diary text.
}
\examples{
wordlists <- get_sentiment_words()
wordlists$positive
wordlists$negative
}
