% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_text.R
\name{load_and_preprocess_text}
\alias{load_and_preprocess_text}
\title{Load text}
\usage{
load_and_preprocess_text(filepath)
}
\arguments{
\item{filepath}{A string. The path to the .txt file.}
}
\value{
A character vector of words.
}
\description{
This function loads a .txt file using read.delim, transforms all text to lowercase,
and splits it into individual words, removing basic stop characters.
}
