% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_sentiment_words.R
\name{match_sentiment_words}
\alias{match_sentiment_words}
\title{Match sentiment words}
\usage{
match_sentiment_words(word_vector, sentiment_vector)
}
\arguments{
\item{word_vector}{Words from the text}

\item{sentiment_vector}{Positive or negative words with optional "*"}
}
\value{
Number of matched words
}
\description{
Match sentiment words
}
