% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_sentiment.R
\name{summarize_sentiment}
\alias{summarize_sentiment}
\title{Summarize sentiment results}
\usage{
summarize_sentiment(word_vector)
}
\arguments{
\item{word_vector}{A character vector of words}
}
\value{
A list with positive, negative counts and ratio
}
\description{
Summarize sentiment results
}
